pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ha-ops-mcp"
version = "0.32.2"
description = "MCP server for Home Assistant operations — database, config, dashboards, entities, and system health"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Maciej Rohleder", email = "dude84@users.noreply.github.com" },
]
keywords = ["home-assistant", "mcp", "model-context-protocol", "smart-home", "ops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]

dependencies = [
    "mcp>=1.0.0",
    "aiohttp>=3.9",
    "websockets>=12.0",
    "ruamel.yaml>=0.18",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "aiomysql>=0.2",
    "asyncpg>=0.29",
    "deepdiff>=7.0",
    "rapidfuzz>=3.6",
    "jsonpatch>=1.33",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
ha-ops-mcp = "ha_ops_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/dude84/ha-ops-mcp"
Repository = "https://github.com/dude84/ha-ops-mcp"
Issues = "https://github.com/dude84/ha-ops-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ha_ops_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["TC003"]  # Path is used at runtime with from __future__ import annotations

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true

# jsonpatch and its jsonpointer dependency have no bundled type stubs.
# The surface we use is small and well-documented (JsonPatch, apply,
# the exception hierarchy), so treating them as Any here is safer than
# wedging PEP 561 compliance onto a third-party dep.
[[tool.mypy.overrides]]
module = ["jsonpatch"]
ignore_missing_imports = true
